scripts/requirements.txt
python-dotenv==1.2.1
requests==2.32.5
requests-html==0.10.0
lxml_html_clean==0.4.3